requirements.txt
# Web Framework
fastapi==0.114.0
uvicorn==0.23.2

# LiveKit & Agent Framework
livekit==1.0.12
livekit-agents==1.2.8
livekit-plugins-openai==1.2.8
livekit-plugins-elevenlabs==1.2.8
livekit-plugins-deepgram==1.2.8
livekit-plugins-silero==1.2.8
livekit-plugins-turn-detector==1.2.8

# AI Services
openai>=1.3.0
elevenlabs==2.9.0
deepgram-sdk==3.11.0

# Configuration
python-dotenv==1.0.1

# Twilio
twilio==9.8.0

# langchain
langchain==0.3.27 
langchain-community==0.3.29 
